# Conflicts:
#	pom.xml

+++ form/service/src/main/java/org/opensingular/form/persistence/dao/AttachmentDao.java
@@ -24,19 +24,22 @@ import javax.inject.Inject;
 import javax.transaction.Transactional;
 
 import org.hibernate.Criteria;
-
-import org.opensingular.lib.commons.base.SingularException;
-import org.opensingular.form.persistence.entity.AttachmentContentEntitty;
+import org.hibernate.Hibernate;
+import org.hibernate.Query;
+import org.joda.time.DateTime;
+import org.opensingular.form.persistence.entity.AbstractFormAttachmentEntity;
+import org.opensingular.form.persistence.entity.AttachmentContentEntity;
 import org.opensingular.form.persistence.entity.AttachmentEntity;
+import org.opensingular.lib.commons.base.SingularException;
 import org.opensingular.lib.support.persistence.BaseDAO;
 
 @SuppressWarnings("serial")
 @Transactional(Transactional.TxType.MANDATORY)
-public class AttachmentDao<T extends AttachmentEntity, C extends AttachmentContentEntitty> extends BaseDAO<T, Long> {
+public class AttachmentDao<T extends AttachmentEntity, C extends AttachmentContentEntity> extends BaseDAO<T, Long> {
 
     @Inject
     private AttachmentContentDao<C> attachmentContentDao;
-    
+
     public AttachmentDao() {
         super((Class<T>) AttachmentEntity.class);
     }
@@ -51,15 +54,21 @@ public class AttachmentDao<T extends AttachmentEntity, C extends AttachmentConte
     }
 
     public T insert(InputStream is, long length, String name){
-        C content = attachmentContentDao.insert(is, length);
+        return insert(is, length, name, null);
+    }
+
+    public T insert(InputStream is, long length, String name, String hashSha1){
+        C content = attachmentContentDao.insert(is, length, hashSha1);
         return insert(createAttachment(content, name));
     }
 
     public void delete(Long id) {
-        T t = get(id);
-        Long codContent = t.getCodContent();
-        delete(t);
-        attachmentContentDao.delete(codContent);
+        final T t = get(id);
+        if (t != null) {
+            Long codContent = t.getCodContent();
+            delete(t);
+            attachmentContentDao.delete(codContent);
+        }
     }
 
     public List<T> list() {
@@ -68,9 +77,9 @@ public class AttachmentDao<T extends AttachmentEntity, C extends AttachmentConte
     }
 
     protected T createAttachment(C content, String name) {
-        
+
         T fileEntity = createInstance();
-        
+
         fileEntity.setCodContent(content.getCod());
         fileEntity.setHashSha1(content.getHashSha1());
         fileEntity.setSize(content.getSize());
@@ -87,4 +96,27 @@ public class AttachmentDao<T extends AttachmentEntity, C extends AttachmentConte
         }
     }
 
+    @SuppressWarnings("unchecked")
+    public List<AttachmentEntity> listOldOrphanAttachments() {
+        StringBuilder hql = new StringBuilder();
+        hql.append(" SELECT a FROM " + AttachmentEntity.class.getName() + " as a ");
+        hql.append(" WHERE a.creationDate < :ontem ");
+        hql.append(" AND NOT EXISTS ( ");
+        hql.append("    SELECT 1 FROM " + AbstractFormAttachmentEntity.class.getName() + " as fa ");
+        hql.append("    WHERE fa.cod.attachmentCod = a.cod ");
+        hql.append(" ) ");
+
+        Query query = getSession().createQuery(hql.toString());
+        Date ontem = new DateTime().minusDays(1).toDate();
+        query.setParameter("ontem", ontem);
+
+        return query.list();
+    }
+
+    @Override
+    public T find(Long aLong) {
+        T t = super.find(aLong);
+        Hibernate.initialize(t);
+        return t;
+    }
 }
